+++ lld/docs/ReleaseNotes.rst
@@ -35,6 +35,7 @@ Breaking changes
 
 * One-dash form of some long option (``--thinlto-*``, ``--lto-*``, ``--shuffle-sections=``)
   are no longer supported.
+* ``--export-dynamic-symbol`` no longer implies ``-u``.
 
 COFF Improvements
 -----------------